+++ include/importmanager.h
@@ -1,15 +1,21 @@
-#ifndef FILEMANAGER_H
+#ifndef IMPORTMANAGER_H
 
-#define FILEMANAGER_H
+#define IMPORTMANAGER_H
 
 #include "atom.h"
+#include "exception.h"
 #include <vector>
 #include <string>
+#include <map>
+#include <utility>
 
-// The file manager is an external component that compartmentalises the code used for
+// The import manager is an external component that compartmentalises the code used for
 // importing data. It works for the Model class.
 namespace ImportMngr{
+  typedef std::array<std::array<double,3>,3> MatR3;
   typedef std::pair<std::string,std::array<double,3>> SymbolPositionPair;
+  typedef std::pair<std::vector<int>,std::vector<double>> SymMatData;
+  typedef std::map<std::string,std::vector<std::string>> AtomDataTable;
 
   // Return struct
   struct UnitCell{
@@ -27,10 +33,13 @@ namespace ImportMngr{
   // Main functions
   std::vector<Atom> readFileXYZ(const std::string&);
   std::pair<std::vector<Atom>,UnitCell> readFilePDB(const std::string&, bool);
-
+  std::pair<std::vector<Atom>,UnitCell> readFileCIF(const std::string&);
+  
   // Aux functions
   std::string strToValidSymbol(std::string str);
   std::vector<std::string> splitLine(const std::string& line);
+  SymMatData convertCifSymmetryElements(const std::vector<std::string>&);
+  std::pair<bool,std::vector<Atom>> convertCifAtomsList(const AtomDataTable&, const MatR3&);
 }
 
 #endif
